Split serializable function code into smaller utility library
This removes a dependency of hpx_functional on hpx_serialization
which causes problems if customization point objects are needed
at lower layers of the software stack.

No new functionality or features are introduced and classes
are simply moved from one sub-dir to another to satisfy
compilation etc.
