+++ libs/core/serializable_function/README.rst
@@ -1,0 +1,16 @@
+
+..
+    Copyright (c) 2019 The STE||AR-Group
+
+    SPDX-License-Identifier: BSL-1.0
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+===
+serializable_function
+===
+
+This library is part of HPX.
+
+Documentation can be found `here
+<https://hpx-docs.stellar-group.org/latest/html/libs/serializable_function/docs/index.html>`__.
