+++ libs/core/serializable_function/docs/index.rst
@@ -1,0 +1,31 @@
+..
+    Copyright (c) 2019 The STE||AR-Group
+
+    SPDX-License-Identifier: BSL-1.0
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+.. _modules_serializable_function:
+
+==========
+serializable_function
+==========
+
+This module provides function wrappers to support serialization of functions and their arguments.
+
+* :cpp:class:`hpx::util::function`
+* :cpp:class:`hpx::util::function_ref`
+* :cpp:class:`hpx::util::unique_function`
+* :cpp:func:`hpx::util::bind`
+* :cpp:func:`hpx::util::bind_back`
+* :cpp:func:`hpx::util::bind_front`
+* :cpp:func:`hpx::util::deferred_call`
+* :cpp:func:`hpx::util::invoke`
+* :cpp:func:`hpx::util::invoke_fused`
+* :cpp:func:`hpx::util::mem_fn`
+* :cpp:func:`hpx::util::one_shot`
+* :cpp:func:`hpx::util::protect`
+* :cpp:class:`hpx::util::result_of`
+
+See the :ref:`API reference <modules_serializable_function_api>` of the module for more
+details.
