+++ libs/core/serializable_function/examples/CMakeLists.txt
@@ -1,0 +1,18 @@
+# Copyright (c) 2019 The STE||AR-Group
+#
+# SPDX-License-Identifier: BSL-1.0
+# Distributed under the Boost Software License, Version 1.0. (See accompanying
+# file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+if(HPX_WITH_EXAMPLES)
+  add_hpx_pseudo_target(examples.modules.serializable_function)
+  add_hpx_pseudo_dependencies(
+    examples.modules examples.modules.serializable_function
+  )
+  if(HPX_WITH_TESTS AND HPX_WITH_TESTS_EXAMPLES)
+    add_hpx_pseudo_target(tests.examples.modules.serializable_function)
+    add_hpx_pseudo_dependencies(
+      tests.examples.modules tests.examples.modules.serializable_function
+    )
+  endif()
+endif()
