+++ libs/core/serializable_function/include/hpx/serializable_function/serialization/detail/vtable/serializable_function_vtable.hpp
@@ -12,7 +12,7 @@
 #include <hpx/functional/detail/empty_function.hpp>
 #include <hpx/functional/detail/function_registration.hpp>
 #include <hpx/functional/detail/vtable/vtable.hpp>
-#include <hpx/functional/serialization/detail/vtable/serializable_vtable.hpp>
+#include <hpx/serializable_function/serialization/detail/vtable/serializable_vtable.hpp>
 #include <hpx/serialization/detail/polymorphic_intrusive_factory.hpp>
 
 #include <string>
