+++ libs/core/serializable_function/include/hpx/serializable_function/serialization/serializable_unique_function.hpp
@@ -10,5 +10,5 @@
 #pragma once
 
 #include <hpx/config.hpp>
-#include <hpx/functional/serialization/detail/serializable_basic_function.hpp>
 #include <hpx/functional/unique_function.hpp>
+#include <hpx/serializable_function/serialization/detail/serializable_basic_function.hpp>
