+++ libs/full/async_distributed/include/hpx/async_distributed/continuation.hpp
@@ -16,12 +16,12 @@
 #include <hpx/async_distributed/continuation_fwd.hpp>
 #include <hpx/async_distributed/trigger_lco_fwd.hpp>
 #include <hpx/components_base/agas_interface.hpp>
-#include <hpx/functional/serialization/serializable_unique_function.hpp>
 #include <hpx/futures/traits/future_traits.hpp>
 #include <hpx/modules/errors.hpp>
 #include <hpx/modules/logging.hpp>
 #include <hpx/naming_base/id_type.hpp>
 #include <hpx/preprocessor/stringize.hpp>
+#include <hpx/serializable_function/serialization/serializable_unique_function.hpp>
 #include <hpx/serialization/base_object.hpp>
 #include <hpx/serialization/serialize.hpp>
