+++ tests/unit/util/function/function_object_size.cpp
@@ -8,64 +8,53 @@
 
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
-#include <hpx/functional/serialization/serializable_function.hpp>
+#include <hpx/serializable_function/serialization/serializable_function.hpp>
 
 #include <hpx/serialization/access.hpp>
 
 #include <cstdint>
 #include <iostream>
 
-using hpx::program_options::variables_map;
 using hpx::program_options::options_description;
+using hpx::program_options::variables_map;
 
 using hpx::util::function;
 
 ///////////////////////////////////////////////////////////////////////////////
 struct small_object
 {
-  private:
+private:
     std::uint64_t x_;
 
     friend class hpx::serialization::access;
 
-    template <
-        typename Archive
-    >
-    void serialize(
-        Archive& ar
-      , unsigned const
-        )
+    template <typename Archive>
+    void serialize(Archive& ar, unsigned const)
     {
-        ar & x_;
+        ar& x_;
         std::cout << "small_object: serialize(" << x_ << ")\n";
     }
 
-  public:
+public:
     small_object()
       : x_(0)
     {
         std::cout << "small_object: default ctor\n";
     }
 
-    small_object(
-        std::uint64_t x
-        )
+    small_object(std::uint64_t x)
       : x_(x)
     {
         std::cout << "small_object: ctor(" << x << ")\n";
     }
 
-    small_object(
-        small_object const& o
-        )
+    small_object(small_object const& o)
       : x_(o.x_)
     {
         std::cout << "small_object: copy(" << o.x_ << ")\n";
     }
 
-    small_object& operator=(
-        small_object const& o
-        )
+    small_object& operator=(small_object const& o)
     {
         x_ = o.x_;
         std::cout << "small_object: assign(" << o.x_ << ")\n";
@@ -77,9 +66,7 @@ struct small_object
         std::cout << "small_object: dtor(" << x_ << ")\n";
     }
 
-    std::uint64_t operator()(
-        std::uint64_t const& z_
-        )
+    std::uint64_t operator()(std::uint64_t const& z_)
     {
         std::cout << "small_object: call(" << x_ << ", " << z_ << ")\n";
         return x_ + z_;
@@ -89,26 +76,21 @@ struct small_object
 ///////////////////////////////////////////////////////////////////////////////
 struct big_object
 {
-  private:
+private:
     std::uint64_t x_;
     std::uint64_t y_;
 
     friend class hpx::serialization::access;
 
-    template <
-        typename Archive
-    >
-    void serialize(
-        Archive& ar
-      , unsigned const
-        )
+    template <typename Archive>
+    void serialize(Archive& ar, unsigned const)
     {
-        ar & x_;
-        ar & y_;
+        ar& x_;
+        ar& y_;
         std::cout << "big_object: serialize(" << x_ << ", " << y_ << ")\n";
     }
 
-  public:
+public:
     big_object()
       : x_(0)
       , y_(0)
@@ -116,28 +98,21 @@ struct big_object
         std::cout << "big_object: default ctor\n";
     }
 
-    big_object(
-        std::uint64_t x
-      , std::uint64_t y
-        )
+    big_object(std::uint64_t x, std::uint64_t y)
       : x_(x)
       , y_(y)
     {
         std::cout << "big_object: ctor(" << x << ", " << y << ")\n";
     }
 
-    big_object(
-        big_object const& o
-        )
+    big_object(big_object const& o)
       : x_(o.x_)
       , y_(o.y_)
     {
         std::cout << "big_object: copy(" << o.x_ << ", " << o.y_ << ")\n";
     }
 
-    big_object& operator=(
-        big_object const& o
-        )
+    big_object& operator=(big_object const& o)
     {
         x_ = o.x_;
         y_ = o.y_;
@@ -150,13 +125,10 @@ struct big_object
         std::cout << "big_object: dtor(" << x_ << ", " << y_ << ")\n";
     }
 
-    std::uint64_t operator()(
-        std::uint64_t const& z_
-      , std::uint64_t const& w_
-        )
+    std::uint64_t operator()(std::uint64_t const& z_, std::uint64_t const& w_)
     {
-        std::cout << "big_object: call(" << x_ << ", " << y_
-                  << z_ << ", " << w_ << ")\n";
+        std::cout << "big_object: call(" << x_ << ", " << y_ << z_ << ", " << w_
+                  << ")\n";
         return x_ + y_ + z_ + w_;
     }
 };
@@ -168,107 +140,107 @@ struct foo
 
 template <typename Archive>
 void serialize(Archive&, foo&, unsigned)
-{}
+{
+}
 
 ///////////////////////////////////////////////////////////////////////////////
 int hpx_main()
 {
+    {{if (sizeof(small_object) <= hpx::util::detail::function_storage_size)
+            std::cout
+        << "object is small\n";
+    else std::cout << "object is large\n";
+
+    small_object const f(17);
+
+    function<std::uint64_t(std::uint64_t const&)> f0(f);
+
+    function<std::uint64_t(std::uint64_t const&)> f1(f0);
+
+    function<std::uint64_t(std::uint64_t const&)> f2;
+
+    f2 = f0;
+
+    f0(7);
+    f1(9);
+    f2(11);
+}
+
+{
+    if (sizeof(big_object) <= hpx::util::detail::function_storage_size)
+        std::cout << "object is small\n";
+    else
+        std::cout << "object is large\n";
+
+    big_object const f(5, 12);
+
+    function<std::uint64_t(std::uint64_t const&, std::uint64_t const&)> f0(f);
+
+    function<std::uint64_t(std::uint64_t const&, std::uint64_t const&)> f1(f0);
+
+    function<std::uint64_t(std::uint64_t const&, std::uint64_t const&)> f2;
+
+    f2 = f0;
+
+    f0(0, 1);
+    f1(1, 0);
+    f2(1, 1);
+}
+}
+// non serializable version
+{
     {
-        {
-            if (sizeof(small_object) <= hpx::util::detail::function_storage_size)
-                std::cout << "object is small\n";
-            else
-                std::cout << "object is large\n";
+        if (sizeof(small_object) <= hpx::util::detail::function_storage_size)
+            std::cout << "object is small\n";
+        else
+            std::cout << "object is large\n";
 
-            small_object const f(17);
+        small_object const f(17);
 
-            function<std::uint64_t(std::uint64_t const&)> f0(f);
+        function<std::uint64_t(std::uint64_t const&), false> f0(f);
 
-            function<std::uint64_t(std::uint64_t const&)> f1(f0);
+        function<std::uint64_t(std::uint64_t const&), false> f1(f0);
 
-            function<std::uint64_t(std::uint64_t const&)> f2;
+        function<std::uint64_t(std::uint64_t const&), false> f2;
 
-            f2 = f0;
+        f2 = f0;
 
-            f0(7);
-            f1(9);
-            f2(11);
-        }
-
-        {
-            if (sizeof(big_object) <= hpx::util::detail::function_storage_size)
-                std::cout << "object is small\n";
-            else
-                std::cout << "object is large\n";
-
-            big_object const f(5, 12);
-
-            function<std::uint64_t(std::uint64_t const&, std::uint64_t const&)>
-                f0(f);
-
-            function<std::uint64_t(std::uint64_t const&, std::uint64_t const&)>
-                f1(f0);
-
-            function<std::uint64_t(std::uint64_t const&, std::uint64_t const&)> f2;
-
-            f2 = f0;
-
-            f0(0, 1);
-            f1(1, 0);
-            f2(1, 1);
-        }
-    }
-    // non serializable version
-    {
-        {
-            if (sizeof(small_object) <= hpx::util::detail::function_storage_size)
-                std::cout << "object is small\n";
-            else
-                std::cout << "object is large\n";
-
-            small_object const f(17);
-
-            function<std::uint64_t(std::uint64_t const&), false> f0(f);
-
-            function<std::uint64_t(std::uint64_t const&), false> f1(f0);
-
-            function<std::uint64_t(std::uint64_t const&), false> f2;
-
-            f2 = f0;
-
-            f0(2);
-            f1(4);
-            f2(6);
-        }
-
-        {
-            if (sizeof(big_object) <= hpx::util::detail::function_storage_size)
-                std::cout << "object is small\n";
-            else
-                std::cout << "object is large\n";
-
-            big_object const f(5, 12);
-
-            function<std::uint64_t(std::uint64_t const&, std::uint64_t const&),
-                false> f0(f);
-
-            function<std::uint64_t(std::uint64_t const&, std::uint64_t const&),
-                false> f1(f0);
-
-            function<std::uint64_t(std::uint64_t const&, std::uint64_t const&),
-                false> f2;
-
-            f2 = f0;
-
-            f0(3, 4);
-            f1(5, 6);
-            f2(7, 8);
-        }
+        f0(2);
+        f1(4);
+        f2(6);
     }
 
-    hpx::finalize();
+    {
+        if (sizeof(big_object) <= hpx::util::detail::function_storage_size)
+            std::cout << "object is small\n";
+        else
+            std::cout << "object is large\n";
 
-    return 0;
+        big_object const f(5, 12);
+
+        function<std::uint64_t(std::uint64_t const&, std::uint64_t const&),
+            false>
+            f0(f);
+
+        function<std::uint64_t(std::uint64_t const&, std::uint64_t const&),
+            false>
+            f1(f0);
+
+        function<std::uint64_t(std::uint64_t const&, std::uint64_t const&),
+            false>
+            f2;
+
+        f2 = f0;
+
+        f0(3, 4);
+        f1(5, 6);
+        f2(7, 8);
+    }
+}
+
+hpx::finalize();
+
+return 0;
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -277,4 +249,3 @@ int main(int argc, char* argv[])
     // Initialize and run HPX
     return hpx::init(argc, argv);
 }
-
